Rebase onto master + adapt to ESM toolchain (#10)
- Convert scripts/state.js and scripts/boot.js from CommonJS to ESM
  (package.json "type":"module" requires .js files to be ESM)
- Convert test/state.test.js to ESM imports; move to tests/state/state.test.js
  to match the project test layout picked up by vitest.config.js
- Remove package-lock.json (project uses pnpm), vitest.config.mjs (superseded
  by vitest.config.js), and test/state.test.js (replaced by tests/state/)
- Export state module from scripts/esm-entry.js so AMD code can require('state')
  via the Vite IIFE + AMD bridge

All 55 tests pass (23 new state tests + 32 existing).
